requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
av==12.1.0
Babel==2.14.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.2.1
config==0.5.1
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.25.1
distro==1.7.0
dnspython==2.6.1
einops==0.8.0
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.0
fqdn==1.5.1
fsspec==2024.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gradio==4.32.2
gradio_client==0.17.0
grpcio==1.64.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.23.2
humanfriendly==10.0
idna==3.6
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
iopath==0.1.10
ipykernel==6.29.0
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-archive==3.4.0
jupyter-events==0.9.0
jupyter-highlight-selected-word==0.2.0
jupyter-lsp==2.2.2
jupyter-nbextensions-configurator==0.6.3
jupyter_client==7.4.9
jupyter_contrib_core==0.4.2
jupyter_contrib_nbextensions==0.7.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lmdb==1.4.1
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
more-itertools==8.10.0
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==6.5.5
notebook_shim==0.2.3
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
omegaconf==2.3.0
onnxruntime==1.18.0
opencv-python==4.9.0.80
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.33.0
regex==2024.5.15
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.17.1
ruff==0.4.7
safetensors==0.4.3
scikit-image==0.23.2
scipy==1.13.1
SecretStorage==3.3.1
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tb-nightly==2.17.0a20240602
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.0
tifffile==2024.5.22
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
transformers==4.41.2
triton==2.2.0
typer==0.12.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.30.1
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.22.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==3.0.3
widgetsnbextension==4.0.9
yacs==0.1.8
yapf==0.40.2
zipp==1.0.0